Add quilt USE flag to dev-util/git
Previously, the git ebuild showed a message about installing quilt if
the user wanted to use git-quiltimport, but took no further action.

Create a USE flag that can depend on quilt for this functionality, and
make the ebuild show the message when quilt is not selected.

If quilt is not set, the message should be used as a warning, that
git-quiltimport will not work.