fix to perl-module.eclass to avoid breaking stage1
This was caused by a dependency on "dev-lang/perl[-build]" for
all perl modules. However, perl modules get merged during stage1
when USE="build" is set. So the dependency has been changed
to "dev-lang/perl". "build" is more of a utility USE variable
used for building stages and it should not enabled on a fully-
built Gentoo system.