change profiles/repo_name from funtoo -> gen3
when attempting to run with paludis as of 0.40.1,
it warns:
paludis@1253288126: [WARNING repository.blacklisted] Repository 'funtoo' is
blacklisted with reason 'Upstream writes code to work with Portage, not to
follow the specification. We can't support this.'.

Since this fork aims to be compatible with both paludis and portage, we
will change the name.