add app-office/unoconv
command line converter for office documents using open office as a backend.

ebuild ganked from graaff overlay with additional patch for working with
openoffice 3.1

However, it is masked because it does not load the openoffice C Extention
properly at least with openoffice-bin; will try it with openoffice compiled from
source.