Fix gcc-code-assist.
No longer use toolchain.eclass.  The libstdc++ includes were not being found
along with other problems.  Now with do a fairly normal ebuild with a blocker on
the specific version gcc-code-assist was branched from for safety.  In this case
it is 4.4.4.  I use it with gcc-4.4.3, and it works fine.